backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-ratelimit==4.1.0

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4

# AI/ML
openai>=1.0.0
numpy==1.24.3

# Local embeddings (no API calls)
sentence-transformers==2.7.0
torch==2.4.1

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
